Extract message handling functions
